# openapi.yaml
openapi: 3.1.0
...
info:
  title: foldbase
  version: "1.0.0-draft.1"
  description: |
    The append-only event log over HTTP, plus its materialized read models and
    their generic query surface: **the log, the log's views, and the way to
    query them.**

    Deployment model: **no shared central instance.** Every app deploys its own
    foldbase as a private sibling in its bundle. The platform's own instance
    serves the platform only.

    ## Auth model

    Mode is declared explicitly via `FOLDBASE_AUTH` (fail-closed: an
    unauthenticated service in production must be opted into, never inferred):

    | `FOLDBASE_AUTH` | Accepted callers |
    |---|---|
    | `none`        | No token. Tenant from `X-Tenant-ID`. Dev / trusted internal networks only. |
    | `service-jwt` | Service tokens only (`type: "service"` claim). User tokens rejected. |
    | `user-jwt`    | Service tokens (full) **and** user tokens (data-plane query only). |

    Token types (same HS256 realm secret; distinguished by the `type` claim):

    | | Tenant source | `X-Tenant-ID` | `X-Auth-*` headers | Control-plane |
    |---|---|---|---|---|
    | **service token** (`type: "service"`) | `X-Tenant-ID` header — **required**, no default (400 if absent). If the token carries `org_id`, tenant is pinned to it and a mismatching header is a 403. | trusted | trusted (forwarded end-user identity) | allowed |
...

    ## Planes

    Every operation is tagged `x-plane: data` or `x-plane: control`.
...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/streams/{streamId}:
    parameters:
      - $ref: "#/components/parameters/StreamId"
      - $ref: "#/components/parameters/TenantHeader"
    post:
      operationId: appendToStream
      tags: [streams]
      x-plane: data
      x-requires-token: service
      summary: Append events to a stream (optimistic concurrency).
      description: |
        **Log first, always.** The append commits before the read-model fold
        runs; a fold failure never fails the append — the response's
        `projected: false` signals a stale projection, repaired by
        `POST /admin/rebuild`.

        `expectedVersion` is the version the caller believes the stream is at
        (0 for a new stream). On mismatch → 409 with the actual version:
        re-read, rebase, retry.

        Requires a **service token** (user tokens are query-only in v1).
        The service stamps the verified token subject into each stored event's
        `metadata` for audit.
      parameters:
        - $ref: "#/components/parameters/AuthUid"
        - $ref: "#/components/parameters/AuthRole"
        - $ref: "#/components/parameters/AuthEmail"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppendRequest" }
      responses:
        "200":
          description: Events durably appended.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppendResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Optimistic-concurrency conflict.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConcurrencyConflict" }
    get:
      operationId: readStream
      tags: [streams]
      x-plane: data
      summary: Read one stream's history, in version order (paged).
      parameters:
        - name: fromVersion
          in: query
          description: Return only events with version > fromVersion.
          schema: { type: integer, minimum: 0 }
        - $ref: "#/components/parameters/ReadLimit"
      responses:
        "200":
          description: The stream's events (empty array for an unknown stream).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StoredEvent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/events:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
    get:
      operationId: readAllEvents
      tags: [streams]
      x-plane: data
      summary: Read the tenant log across streams, in global order (paged).
      description: >
        Cursor pagination: fromGlobalSeq + limit (default 1000, max 10000).
        `type` narrows to one stream category — a read-model-free category read.
      parameters:
        - name: fromGlobalSeq
          in: query
          description: Return only events with globalSeq > fromGlobalSeq (replay cursor).
          schema: { type: integer, minimum: 0 }
        - name: type
          in: query
          description: Only events of streams with this streamType (ADR-008).
          schema: { $ref: "#/components/schemas/Identifier" }
        - $ref: "#/components/parameters/ReadLimit"
      responses:
        "200":
          description: The tenant's events in globalSeq order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StoredEvent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/events/by-correlation/{correlationId}:
    parameters:
      - name: correlationId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
      - $ref: "#/components/parameters/TenantHeader"
    get:
      operationId: readByCorrelation
      tags: [streams]
      x-plane: data
      summary: Read every event carrying a correlationId, in global order (lineage / tracing).
      parameters:
        - $ref: "#/components/parameters/ReadLimit"
      responses:
        "200":
          description: Correlated events in globalSeq order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StoredEvent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/subscribe:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
    get:
      operationId: subscribe
      tags: [streams]
      x-plane: data
      x-requires-token: service
      summary: Realtime event stream over Server-Sent Events (ADR-009).
      description: |
        Streams appended events as `text/event-stream`. **Service-level callers
        only** — the raw log bypasses row policies, so the owning app subscribes
        with a service token and relays to its own users.

        Delivery is ordered and gap-free: the server first replays the log from
        the cursor (`Last-Event-ID` header on reconnect, else `fromGlobalSeq`),
        then tails live appends from an in-process bus. Each SSE frame carries
        `id: <globalSeq>` (the reconnect cursor) and `event: <EventType>`; `data`
        is the StoredEvent JSON. `: ping` comment lines are heartbeats. A
        subscriber that falls too far behind is dropped and should reconnect
        (its cursor makes catch-up lossless).
      parameters:
        - name: type
          in: query
          description: Only events of streams with this streamType (category subscription).
          schema: { $ref: "#/components/schemas/Identifier" }
        - name: fromGlobalSeq
          in: query
          description: Start cursor (superseded by the Last-Event-ID header on reconnect).
          schema: { type: integer, minimum: 0 }
        - name: Last-Event-ID
          in: header
          description: SSE reconnect cursor — the last globalSeq delivered.
          schema: { type: string }
      responses:
        "200":
          description: An open SSE stream.
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/query/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: A registered projection name. Physical tables are never addressable.
        schema: { $ref: "#/components/schemas/Identifier" }
      - $ref: "#/components/parameters/TenantHeader"
    post:
      operationId: query
      tags: [query]
      x-plane: data
      summary: Query a registered read model. POST-only, by design.
      description: |
        The single generic read endpoint — there are **no per-table routes and
        no GET variant**; the JSON body is the extension point.

        Row access: deny-by-default select policies. With a **service token**,
        the auth context binds from the forwarded `X-Auth-*` headers; with a
        **user token**, it binds from the verified token claims and `X-Auth-*`
        headers are ignored. A policy referencing an `:auth_*` value absent
        from the context denies (403) rather than widening.
      parameters:
        - $ref: "#/components/parameters/AuthUid"
        - $ref: "#/components/parameters/AuthRole"
        - $ref: "#/components/parameters/AuthEmail"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueryRequest" }
      responses:
        "200":
          description: Matching rows.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueryResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/projections:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
    put:
      operationId: putProjection
      tags: [definitions]
      x-plane: control
      x-requires-token: service
      summary: Register or update a projection definition (idempotent, auto-rebuild).
      description: |
        Definitions are DATA authored by the owning app and pushed on ITS boot,
        like migrations. Registration creates/evolves the `read_<name>` table
        (additive ALTER only) and always ends with a rebuild for the calling
        tenant — "events first, definition later" is safe; ordering never
        matters.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectionDef" }
      responses:
        "200":
          description: Definition saved; read model rebuilt for the calling tenant.
          content:
            application/json:
              schema:
                type: object
                required: [ok, name, rebuiltFrom]
                properties:
                  ok: { type: boolean, const: true }
                  name: { $ref: "#/components/schemas/Identifier" }
                  rebuiltFrom:
                    type: integer
                    minimum: 0
                    description: Number of events replayed into the rebuilt model.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/policies:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
    put:
      operationId: putPolicy
      tags: [definitions]
      x-plane: control
      x-requires-token: service
      summary: Register or update a select policy (idempotent).
      description: |
        Deny-by-default — a projection with no policy for the caller's role
        (nor a `*` fallback) is not queryable at all. There is no bypass: a
        service reading without end-user context needs an explicit
        `role: "service"` policy.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyDef" }
      responses:
        "200":
          description: Policy saved.
          content:
            application/json:
              schema:
                type: object
                required: [ok, name, role]
                properties:
                  ok: { type: boolean, const: true }
                  name: { $ref: "#/components/schemas/Identifier" }
                  role: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/reload:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
    post:
      operationId: adminReload
      tags: [admin]
      x-plane: control
      x-requires-token: service
      summary: Re-read _projections/_policies into memory.
      responses:
        "200":
          description: Registry reloaded.
          content:
            application/json:
              schema:
                type: object
                required: [ok, projections]
                properties:
                  ok: { type: boolean, const: true }
                  projections:
                    type: array
                    items: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/rebuild:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
    post:
      operationId: adminRebuild
      tags: [admin]
      x-plane: control
      x-requires-token: service
      summary: Replay the tenant log into one projection, or all of them.
      description: |
        Read models are disposable — the log is the truth. Rebuild wipes the
        projection's rows for the calling tenant and refolds the log in
        globalSeq order.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Identifier"
                  description: Rebuild only this projection; omit for all.
      responses:
        "200":
          description: Rebuild complete.
          content:
            application/json:
              schema:
                type: object
                required: [ok, rebuiltFrom]
                properties:
                  ok: { type: boolean, const: true }
                  rebuiltFrom:
                    type: integer
                    minimum: 0
                    description: Number of events replayed.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256, realm secret held by the platform broker — app code never holds
        the signing key. Claims contract:
          `type`   — "service" | "user" (absent ⇒ user)
          `org_id` — user token: the tenant (required);
                     service token: optional narrowing (pins the tenant)
          `sub`    — user token: end-user id (required); service token: app id
          `role`, `email` — user token identity, bound into policies
        Not sent at all in `FOLDBASE_AUTH=none` deployments.
  parameters:
    ReadLimit:
      name: limit
      in: query
      required: false
      description: Page size for log reads. Default 1000; explicit values up to 10000.
      schema: { type: integer, minimum: 1, maximum: 10000, default: 1000 }
    StreamId:
      name: streamId
      in: path
      required: true
      description: The entity/aggregate id whose history this stream is.
      schema: { type: string, minLength: 1 }
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: |
        Tenant selector. Semantics by caller:
        service token → REQUIRED, no default (400 if absent; 403 on mismatch
        with a token-pinned `org_id`). user token → ignored (tenant comes from
        the token). `none` mode → required.
      schema: { type: string, minLength: 1 }
    AuthUid:
      name: X-Auth-UID
      in: header
      required: false
      description: >
        Forwarded end-user id. Honored only with a service token (the caller
        verified the end-user itself); ignored with user tokens.
      schema: { type: string }
    AuthRole:
      name: X-Auth-Role
      in: header
      required: false
      description: Forwarded end-user role. Service tokens only.
      schema: { type: string }
    AuthEmail:
      name: X-Auth-Email
      in: header
      required: false
      description: Forwarded end-user email. Service tokens only.
      schema: { type: string }
  responses:
    BadRequest:
      description: Malformed body, unknown/forbidden column, invalid operator, or missing tenant.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, expired, or unscoped token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >
        No select policy for the caller (deny-by-default), unsatisfiable policy,
        user token on a service-only operation, or tenant/pin mismatch.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Unknown projection name.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Identifier:
      type: string
      pattern: "^[a-z][a-z0-9_]*$"
      description: >
        Lowercase identifier (a-z, 0-9, _). The ONLY shape a name may take
        before it can ever reach an SQL identifier position; also the reason
        `_`-prefixed physical tables are structurally unreachable.
...
    ConcurrencyConflict:
      type: object
      required: [error, actual]
      properties:
        error: { type: string, const: concurrency_conflict }
        actual:
          type: integer
          minimum: 0
          description: The stream's actual current version — re-read from here and retry.
...
    StoredEvent:
      allOf:
        - $ref: "#/components/schemas/NewEvent"
        - type: object
          required: [tenant, id, streamType, version, globalSeq, recordedAt, payload, metadata]
          properties:
            tenant: { type: string, minLength: 1 }
            streamType:
              type: string
              description: The stream's kind ('' = untyped); same value stream-wide.
            id: { type: string, format: uuid }
            version:
              type: integer
              minimum: 1
              description: Position within its stream (1, 2, 3, …).
            globalSeq:
              type: integer
              minimum: 1
              description: Monotonic order across the whole tenant log.
            recordedAt:
              type: integer
              minimum: 0
              description: Epoch-ms instant the fact was recorded.
    AppendRequest:
      type: object
      required: [expectedVersion, events]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
          description: The stream version the caller believes is current (0 = new stream).
        streamType:
          type: string
          pattern: "^([a-z][a-z0-9_]*)?$"
          default: ""
          description: >
            The aggregate's kind (ADR-008). Fixed by the stream's FIRST append;
            a later conflicting non-empty value is a 400. Drives the
            GET /v1/events?type= category read. Optional — omitted/empty means
            an untyped stream (back-compat).
        events:
          type: array
          items: { $ref: "#/components/schemas/NewEvent" }
          description: Appended atomically, versions assigned in order.
    AppendResponse:
      type: object
      required: [events, version, projected]
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/StoredEvent" }
        version:
          type: integer
          minimum: 0
          description: The stream version after this append.
        projected:
          type: boolean
          description: >
            Whether the read-model fold landed. false = the events are durable
            but at least one projection is stale; POST /admin/rebuild repairs.
...
    QueryResponse:
      type: object
      required: [rows, limit, offset]
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        limit: { type: integer }
        offset: { type: integer }
...